Add a Derive for VariantName which provides a variant_name method
This is intentionally different from
`impl From<Enum> for &'static str {...}`

This intends to be more declarative about the meaning of the string.

There are many ways to convert an enum to a string (eg display, debug,
variant name, etc), and as_static() and into() are ambiguous and feel
overloaded for some callers.

Some callers may want to have a more declarative unambiguous name for
a method with this signature fn (&self) -> &'static str.

This provides it! Implementation wise, it's identical to the
deprecated as_static()