Minimal fix to iterator behavior & drop mut on mask_iter().
This is just minimal changes without to fix the iterator impl which
doesn't require changing any of the trait bounds.

I'd thought we couldn't directly call std::mem::size_of::<R>(), but noticed
that it does work. I think a better iter implementation is possible but may require
different trait bounds..

Along with additional tests, and a change to remove mut on mask was useless
since made a copy/clone.